+++ examples/cleanup/playwright.config.ts
@@ -1,11 +1,9 @@
 import { defineConfig } from '@playwright/test';
 import { globalCache } from '@vitalets/global-cache';
 
-export default defineConfig({
-  testDir: './test',
-  globalSetup: globalCache.setup,
-  globalTeardown: [
-    require.resolve('./test/cleanup'), // <-- custom teardown script for cleanup
-    globalCache.teardown,
-  ],
-});
+export default globalCache.playwright(
+  defineConfig({
+    testDir: './test',
+    globalTeardown: require.resolve('./test/cleanup'), // <-- custom teardown script for cleanup
+  }),
+);