+++ CHANGELOG.md
@@ -6,7 +6,8 @@
 * chore: use relative path in sig
 * fix: check TTL within single test run
 * feat: split values per `runId`
-* feat: add `globalCache.teardownReporter`
+* feat: add `globalCache.reporter`
+* feat: add `globalCache.playwright()` helper
 
 ## [0.2.2] - 2025-09-03
 * feat: better signature checks.